refactor: rename achievements to stats worker
