--- packages/workers/achievements/README.md
@@ -1,1 +1,0 @@
-# Achievements worker
