+++ packages/workers/stats/.dev.vars.example
