+++ packages/workers/stats/README.md
@@ -1,0 +1,1 @@
+# Stats worker
