+++ packages/workers/stats/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@workers/achievements",
+  "name": "@workers/stats",
   "version": "0.0.0",
   "private": true,
   "license": "AGPL-3.0",
