+++ packages/workers/stats/src/handlers/streaksCalendar.ts
