+++ packages/workers/stats/src/helpers/filteredNames.ts
