+++ packages/workers/stats/src/index.ts
@@ -3,10 +3,10 @@ import response from '@hey/lib/response';
 import { createCors, error, Router, status } from 'itty-router';
 
 import haveUsedHey from './handlers/haveUsedHey';
+import profileImpressions from './handlers/profileImpressions';
 import publicationViews from './handlers/publicationViews';
 import streaksCalendar from './handlers/streaksCalendar';
 import streaksList from './handlers/streaksList';
-import userImpressions from './handlers/userImpressions';
 import buildRequest from './helpers/buildRequest';
 import type { Env, WorkerRequest } from './types';
 
@@ -22,12 +22,12 @@ router
   .head('*', () => status(200))
   .get('/', (request: WorkerRequest) =>
     response({
-      message: 'gm, to achievements service 👋',
+      message: 'gm, to stats service 👋',
       version: request.env.RELEASE ?? 'unknown'
     })
   )
   .post('/publicationViews', publicationViews)
-  .get('/userImpressions', userImpressions)
+  .get('/profileImpressions', profileImpressions)
   .get('/haveUsedHey/:id', haveUsedHey)
   .get('/streaks/:id', streaksCalendar)
   .get('/streaks/:id/:date', streaksList)
